+++ spec/spec_helper.rb
@@ -10,5 +10,9 @@ RSpec.configure do |config|
 
   config.filter_run :focus => true
   config.run_all_when_everything_filtered = true
+
+  config.after :each do
+    Timecop.return
+  end
 end